feat: add genre functions

+++ src/main/java/ru/klasix12/film_bot/service/impl/FilmBotServiceImpl.java
@@ -18,6 +18,7 @@ public class FilmBotServiceImpl implements FilmBotService {
     private final KinopoiskApiService kinopoiskApiService;
     private final UserService userService;
     private final FilmService filmService;
+    private final Random random = new Random();
 
     public FilmBotServiceImpl(KinopoiskApiService kinopoiskApiService, UserService userService, FilmService filmService) {
         this.kinopoiskApiService = kinopoiskApiService;
@@ -53,42 +54,81 @@ public class FilmBotServiceImpl implements FilmBotService {
                         %s
                         Оценка кинопоиск: %s
                         Оценка IMDb: %s
+                        Жанры: %s
                         Описание: %s""",
                 user.getUsername(),
                 film.get().getNameRu(),
                 film.get().getRatingKinopoisk(),
                 film.get().getRatingImdb(),
+                film.get().getGenres(),
                 film.get().getDescription());
     }
 
     @Override
     public String findAll() {
         List<Film> films = filmService.findAll();
+        if (films.isEmpty()) {
+            return "Список фильмов пуст";
+        }
         StringBuilder sb = new StringBuilder();
 
         for (int i = 0; i < films.size(); i++) {
-            sb.append(i).append(") ").append(films.get(i)).append("\n");
+            sb.append(i + 1).append(") ").append(films.get(i).getNameRu()).append("\n");
         }
         return sb.toString();
     }
 
     @Override
     public String getRandomFilm() {
-        Random random = new Random();
         List<Film> films = filmService.findAll();
+        if (films.size() <= 0) {
+            return "Список фильмов пуст";
+        }
         Film film = films.get(random.nextInt(films.size()));
         filmService.remove(film);
+        return formatFilm(film);
+    }
+
+    @Override
+    public String findAllByGenre(String genre) {
+        List<Film> films = filmService.getFilmsByGenre(genre);
+        if (films.isEmpty()) {
+            return "По вашему запросу ничего не нашлось :(";
+        }
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < films.size(); i++) {
+            sb.append(i + 1).append(") ").append(films.get(i).getNameRu());
+        }
+        return sb.toString();
+    }
+
+    @Override
+    public String getRandomFilmByGenre(String genre) {
+        List<Film> films = filmService.getFilmsByGenre(genre);
+        if (films.isEmpty()) {
+            return "По вашему запросу ничего не нашлось :(";
+        }
+        Film film = films.get(random.nextInt(films.size()));
+        filmService.remove(film);
+        return formatFilm(film);
+    }
+
+    private String formatFilm(Film film) {
         return String.format("""
                         Выпал фильм: %s
                         Оценка кинопоиск: %s
                         Оценка IMDb: %s
+                        Жанры: %s
                         Описание: %s
+                        Ссылка: %s
                         Добавил: @%s
                         """,
                 film.getNameRu(),
                 film.getRatingKinopoisk(),
                 film.getRatingImdb(),
+                film.getGenres(),
                 film.getDescription(),
+                film.getUrl().replace("kino", "ss"),
                 film.getUser().getUsername());
     }
 }
